Move architecture functions and define `Model`
