+++ src/model.jl
@@ -1,0 +1,8 @@
+struct Model
+    arch::AbstractArchitecture
+    mesh::Mesh
+    params::Parameters
+    inversion::InversionToolkit
+    evolution::EvolutionToolkit
+    state::State
+end
