+++ src/nuPGCM.jl
@@ -45,32 +45,7 @@ function set_out_dir!(dir)
     end
 end
 
-# define CPU and GPU architectures (see: Oceananigans.jl/src/Architectures.jl)
-abstract type AbstractArchitecture end
-
-struct CPU <: AbstractArchitecture end
-struct GPU <: AbstractArchitecture end
-
-# convert types from one architecture to another
-on_architecture(::CPU, a::Array) = a
-on_architecture(::GPU, a::Array) = CuArray(a)
-
-on_architecture(::CPU, a::CuArray) = Array(a)
-on_architecture(::GPU, a::CuArray) = a
-
-on_architecture(::CPU, a::SparseMatrixCSC) = a
-on_architecture(::GPU, a::SparseMatrixCSC) = CuSparseMatrixCSR(a)
-
-on_architecture(::CPU, a::CuSparseMatrixCSR) = SparseMatrixCSC(a)
-on_architecture(::GPU, a::CuSparseMatrixCSR) = a
-
-# determine architecture array is on 
-architecture(::Array) = CPU()
-architecture(::CuArray) = GPU()
-architecture(::SparseMatrixCSC) = CPU()
-architecture(::CuSparseMatrixCSR) = GPU()
-
-# physical dimension of the problem
+#TODO: Remove
 abstract type AbstractDimension end
 struct TwoD <: AbstractDimension 
     n::Int
@@ -83,27 +58,33 @@ ThreeD() = ThreeD(3)
 string(::TwoD) = "2D"
 string(::ThreeD) = "3D"
 
+# include all the module code
+include("architectures.jl")
 include("utils.jl")
 include("plotting.jl")
+include("parameters.jl")
 include("spaces.jl")
 include("dofs.jl")
 include("meshes.jl")
 include("matrices.jl")
-include("inversion.jl")
 include("preconditioners.jl")
+include("inversion.jl")
+include("evolution.jl")
 include("state.jl")
+include("model.jl")
 
 export 
 out_dir,
 set_out_dir!,
+AbstractDimension,
+TwoD,
+ThreeD,
+# architectures.jl
 AbstractArchitecture,
 CPU,
 GPU,
 on_architecture,
 architecture,
-AbstractDimension,
-TwoD,
-ThreeD,
 # utils.jl
 chebyshev_nodes,
 hrs_mins_secs,
@@ -118,6 +99,8 @@ plot_slice,
 plot_profiles,
 sim_plots,
 plot_sparsity_pattern,
+# parameters.jl
+Parameters,
 # spaces.jl
 Spaces,
 # dofs.jl
@@ -131,16 +114,20 @@ get_p_to_t,
 ∂y,
 ∂z,
 build_matrices,
+# preconditioners.jl
+mul!,
 # inversion.jl
 InversionToolkit,
 invert!,
-# preconditioners.jl
-mul!,
+# evolution.jl
+EvolutionToolkit,
 # state.jl
 State,
 rest_state,
 set_state!,
 save,
-load_state
+load_state,
+# model.jl
+Model
 
 end # module
