+++ src/parameters.jl
@@ -1,0 +1,5 @@
+struct Parameters
+    ε::Real
+    α::Real
+    μϱ::Real
+end
